@@ -2,9 +2,6 @@
   "name": "hyphy-vision",
   "description": "HyPhy's visual interface",
   "homepage": "http://hyphy.org",
-  "jest": {
-    "config": "jest.config.js"
-  },
   "keywords": [
     "phylogenetics",
     "phylogeny",
@@ -23,44 +20,43 @@
     "url": "git://github.com/veg/hyphy-vision.git"
   },
   "main": "dist/hyphyvision.js",
-  "version": "2.7.0",
+  "version": "2.8.10",
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.12.1",
-    "alignment.js": "0.1.1",
+    "alignment.js": "0.3.0",
     "bootstrap": "4.x",
     "chi-squared": "^1.1.0",
     "circos": "^2.1.0",
     "create-react-class": "^15.6.3",
-    "csvexport": "1.1.0",
     "d3": "3.x",
     "d3-save-svg": "0.0.2",
+    "export-to-csv": "^0.2.1",
     "express": "^4.16.2",
-    "file-saver": "^1.3.3",
+    "file-saver": "^2.0.2",
     "immutable": "3.8.2",
     "in-browser-download": "^1.0.0",
-    "jquery": "1.x",
+    "jquery": "3.x",
     "jquery-ui": "1.x",
     "jquery-ui-bundle": "^1.11.4",
     "jquery-ui-dist": "^1.12.1",
     "json-loader": "^0.5.7",
-    "lodash": "3.x",
+    "lodash": "4.x",
     "lodash.clonedeep": "^4.5.0",
     "phylotree": "0.1.6",
     "popper.js": "^1.14.1",
     "pretty-data": "^0.40.0",
     "prop-types": "^15.5.10",
-    "react": "16",
+    "react": "16.13.1",
     "react-copy-to-clipboard": "^5.0.0",
-    "react-dom": "16",
+    "react-dom": "16.13.1",
     "react-json-view": "^1.14.0",
-    "react-phylotree": "0.1.0",
+    "react-phylotree": "0.4.3",
     "react-range": "^1.5.3",
-    "react-router-dom": "^4.2.2",
-    "react-router-hash-link": "^1.1.1",
-    "react-scrollchor": "^4.2.1",
+    "react-router-dom": "^5.1.2",
+    "react-router-hash-link": "^2.0.0",
+    "react-scrollchor": "^6.0.0",
     "save-svg-as-png": "^1.2.0",
-    "supervisor": "^0.12.0",
-    "underscore": "^1.9.1"
+    "underscore": "1.9"
   },
   "devDependencies": {
     "@babel/core": "^7.8.7",
@@ -78,11 +74,10 @@
     "bootstrap-webpack": "0.0.x",
     "copy-webpack-plugin": "^5.1.1",
     "css-loader": "^3.4.2",
-    "electron": "^1.7.9",
     "enzyme": "3",
     "enzyme-adapter-react-16": "^1.14.0",
-    "eslint": "^6.8.0",
-    "eslint-loader": "^3.0.3",
+    "eslint": "^7.5.0",
+    "eslint-loader": "^4.0.0",
     "eslint-plugin-react": "^7.1.0",
     "exports-loader": "^0.7.0",
     "expose-loader": "^0.7.1",
@@ -90,11 +85,11 @@
     "font-awesome-webpack": "^0.0.5-beta.2",
     "font-awesome-webpack-4": "^1.0.0",
     "html-webpack-plugin": "^4.0.0-alpha.2",
-    "husky": "^0.14.3",
+    "husky": "^4.2.3",
     "imports-loader": "^0.8.0",
     "interpolate-html-plugin": "^3.0.0",
-    "jest": "^25.1.0",
-    "mini-css-extract-plugin": "^0.9.0",
+    "jest": "^26.0.1",
+    "mini-css-extract-plugin": "^0.11.0",
     "node-sass": "^4.8.3",
     "npm-version": "^1.1.0",
     "preload-webpack-plugin": "3.0.0-beta.3",
@@ -115,12 +110,11 @@
     "webpack-eslint-plugin": "^1.2.0"
   },
   "scripts": {
-    "build": "webpack",
+    "build": "webpack --mode production",
     "build:library": "webpack --config webpack.config.library.js",
     "lint": "eslint --ext .js,.jsx --ignore-pattern public,static,dist ./src/",
     "test": "node ./node_modules/jest-cli/bin/jest.js",
     "start": "node_modules/.bin/webpack-dev-server --open",
-    "electron": "webpack && node_modules/.bin/electron electron.js",
     "format": "prettier 'src/**/*.{js,jsx}' --write",
     "precommit": "pretty-quick --staged"
   },
