@@ -4,11 +4,10 @@ const express = require("express"),
 const port = process.env.PORT || 8000;
 const app = express();
 
-app.use(express.static("."));
-app.use(express.static("public"));
+app.use(express.static("dist"));
 
-app.get("/:methods/?", (req, res) => {
-  res.sendFile(path.resolve(__dirname, "public", "index.html"));
+app.get("*", (req, res) => { // lgtm [js/missing-rate-limiting]
+  res.sendFile(path.resolve(__dirname, "dist", "index.html"));
 });
 
 app.listen(port, () => {
