@@ -142,16 +142,17 @@ class MultiHitContents extends React.Component {
     // These variables are to be used for DatamonkeyTable
     const erHeaders = [
       ["Site", "Site Index"],
-      ['<i class="fas fa-dice-three fa-3x"></i>', "Three-hit"],
+      ['<i class="fas fa-dice-three fa-2x"></i> vs. <i class="fas fa-dice-two fa-2x"></i>',
+       "3H vs 2H"],
       [
-        '<i class="fas fa-dice-three fa-2x"></i> vs. <i class="fas fa-dice-two fa-2x"></i>',
-        "Three-hit islands vs 2-hit"
+        '<i class="fas fa-compass fa-2x"></i> vs. <i class="fas fa-dice-two fa-2x"></i>',
+        "3HSI vs 2H"
       ],
       [
-        '<i class="fas fa-compass fa-3x"></i>',
-        "Three-hit vs three-hit islands"
+        '<i class="fas fa-compass fa-2x"></i> vs. <i class="fas fa-dice-three fa-2x"></i>',
+        "3H vs 3HSI"
       ],
-      ['<i class="fas fa-dice-two fa-3x"></i>', "Two-hit"]
+      ['<i class="fas fa-dice-two fa-2x"></i> vs. <i class="fas fa-dice-one fa-2x"></i>', "2H vs 1H"]
     ];
 
     const siteLogLHeaders = [
@@ -230,7 +231,7 @@ class MultiHitContents extends React.Component {
       whichTable: whichTable,
       testResults: [],
       xaxis: "Site",
-      yaxis: "Three-hit",
+      yaxis: "3H",
       xOptions: ["Site"],
       yOptions: [],
       copy_transition: false,
@@ -249,28 +250,28 @@ class MultiHitContents extends React.Component {
       maxTransitionCount: 0,
       showLegend: true,
       erRanges: {
-        "Three-hit": [0, 10],
-        "Three-hit islands vs 2-hit": [0, 10],
-        "Three-hit vs three-hit islands": [0, 10],
-        "Two-hit": [0, 10]
+        "3H vs 2H": [0, 10],
+        "3HSI vs 2H": [0, 10],
+        "3H vs 3HSI": [0, 10],
+        "2H vs 1H": [0, 10]
       },
       ERThresholds: {
-        "Three-hit": [0, 5],
-        "Three-hit islands vs 2-hit": [0, 5],
-        "Three-hit vs three-hit islands": [0, 5],
-        "Two-hit": [0, 5]
+        "3H vs 2H": [0, 5],
+        "3HSI vs 2H": [0, 5],
+        "3H vs 3HSI": [0, 5],
+        "2H vs 1H": [0, 5]
       },
       minERs: {
-        "Three-hit": 0,
-        "Three-hit islands vs 2-hit": 0,
-        "Three-hit vs three-hit islands": 0,
-        "Two-hit": 0
+        "3H vs 2H": 0,
+        "3HSI vs 2H": 0,
+        "3H vs 3HSI": 0,
+        "T2H vs 1H": 0
       },
       maxERs: {
-        "Three-hit": 5,
-        "Three-hit islands vs 2-hit": 5,
-        "Three-hit vs three-hit islands": 5,
-        "Two-hit": 5
+        "T3H vs 2H": 5,
+        "3HSI vs 2H": 5,
+        "3H vs 3HSI": 5,
+        "2H vs 1H": 5
       },
       fits: {}
     };
@@ -876,25 +877,25 @@ class MultiHitContents extends React.Component {
         <h3>Likelihood Test Results</h3>
         <div className="row">
           {this.getTestResultCard(
-            '<i class="fas fa-dice-three"></i> Triple-hit vs single-hit',
+            '<i class="fas fa-dice-three"></i> 3H vs <i class="fas fa-dice-one"></i> 1H',
             this.state.testResults["Triple-hit vs single-hit"]
           )}
           {this.getTestResultCard(
-            '<i class="fas fa-dice-three"></i> vs. <i class="fas fa-dice-two"></i> Triple-hit vs double-hit',
+            '<i class="fas fa-dice-three"></i> 3H vs. <i class="fas fa-dice-two"></i> 2H',
             this.state.testResults["Triple-hit vs double-hit"]
           )}
           {this.getTestResultCard(
-            '<i class="fas fa-compass"></i> Triple-hit vs. Triple-hit-island',
+            '<i class="fas fa-dice-three"></i> 3H vs <i class="fas fa-compass"></i> 3HSI',
             this.state.testResults["Triple-hit vs Triple-hit-island"]
           )}
         </div>
         <div className="row">
           {this.getTestResultCard(
-            '<i class="fas fa-dice-two"></i> Double-hit vs Single-hit Test',
+            '<i class="fas fa-dice-two"></i> 2H vs <i class="fas fa-dice-one"></i> 1H',
             this.state.testResults["Double-hit vs single-hit"]
           )}
           {this.getTestResultCard(
-            "Triple-hit-island vs double-hit",
+            '<i class="fas fa-compass"></i> 3HSI vs <i class="fas fa-dice-two"></i> 2H',
             this.state.testResults["Triple-hit-island vs double-hit"]
           )}
         </div>
@@ -958,8 +959,8 @@ class MultiHitContents extends React.Component {
               See <a href="//hyphy.org/methods/selection-methods/#fel">here</a>{" "}
               for more information about the MultiHit method.
               <br />
-              Please cite PMID{" "}
-              <a href="//www.ncbi.nlm.nih.gov/pubmed/15703242">15703242</a> if
+              Please cite biorxv{" "}
+              <a href="//https://www.biorxiv.org/content/10.1101/2020.05.13.091652v1"> </a> if
               you use this result in a publication, presentation, or other
               scientific work.
             </small>
@@ -1020,16 +1021,16 @@ class MultiHitContents extends React.Component {
         <div id="results">
           <MainResult
             summary_for_rendering={this.getSummaryForRendering()}
-            method_ref="http://hyphy.org/methods/selection-methods/#multi-hit"
-            citation_ref="//www.ncbi.nlm.nih.gov/pubmed/15703242"
-            citation_number="PMID 15703242"
+            method_ref="http://hyphy.org/methods/other/FitMultiModel/"
+            citation_ref="https://www.biorxiv.org/content/10.1101/2020.05.13.091652v1"
+            citation_number="biorxv"
           />
 
-          <div id="table-tab" className="row hyphy-row">
-            <div id="hyphy-mle-fits" className="col-md-12">
+          <div id="site-plot-tab" className="row hyphy-row">
+            <div className="col-md-12">
               <Header
-                title="Model Test Statistics Per Site"
-                popover="<p>Hover over a column header for a description of its content.</p>"
+                title="Model Test Statistics Per Site Plot"
+                popover="<p>Hover over a point to see its values</p>"
               />
 
               <DatamonkeyGraphMenu
@@ -1085,7 +1086,16 @@ class MultiHitContents extends React.Component {
                   </label>
                 </div>
               </div>
-
+            </div>
+           </div>
+           <div id="table-tab" className="row hyphy-row">
+            <div id="hyphy-mle-fits" className="col-md-12">
+             
+              <Header
+                title="Model Test Statistics Per Site Table"
+                popover="<p>Hover over a column header for a description of its content.</p>"
+              />
+              
               <DatamonkeyTable
                 headerData={this.state.siteTableHeaders[this.state.whichTable]}
                 bodyData={this.state.siteTableContent[this.state.whichTable]}
@@ -1246,14 +1256,15 @@ class MultiHitContents extends React.Component {
   }
 }
 
-function MultiHit(props) {
+export function MultiHit(props) {
   return (
     <ResultsPage
       data={props.data}
       scrollSpyInfo={[
         { label: "summary", href: "summary-tab" },
+        { label: "site plot", href: "site-plot-tab" },
         { label: "table", href: "table-tab" },
-        { label: "plot", href: "plot-tab" }
+        { label: "substitution plot", href: "plot-tab" }
       ]}
       methodName="Multi-Hit"
       fasta={props.fasta}
@@ -1265,7 +1276,13 @@ function MultiHit(props) {
   );
 }
 
-function render_multihit(data, element, fasta, originalFile, analysisLog) {
+export default function render_multihit(
+  data,
+  element,
+  fasta,
+  originalFile,
+  analysisLog
+) {
   ReactDOM.render(
     <MultiHit
       data={data}
@@ -1276,5 +1293,3 @@ function render_multihit(data, element, fasta, originalFile, analysisLog) {
     document.getElementById(element)
   );
 }
-
-export { MultiHit, render_multihit };
