@@ -23,7 +23,6 @@ import { saveSvgAsPng } from "save-svg-as-png";
 import Phylotree, { placenodes, phylotreev1 } from "react-phylotree";
 import { SitePlotAxis, fastaParser, colors } from "alignment.js";
 import CodonColumn from "./components/codon_column.jsx";
-import AlignmentTree from "./components/tree-alignment.jsx";
 
 require("../datamonkey/helpers.js");
 
@@ -1153,6 +1152,325 @@ var SLACGraphs = createReactClass({
   }
 });
 
+
+function SLACAlignmentTreeCountWrapper(props) {
+  if (!props.json) return null;
+  const { trees } = props.json,
+    branchAttributes = props.json['branch attributes'],
+    partitions = props.json['data partitions'],
+    number_of_partitions = _.keys(partitions).length,
+    all_partition_data = [],
+    partition_map = [];
+  for(let partition=0; partition < number_of_partitions; partition++) {
+    let partition_data = {},
+      sites_in_partition = partitions[partition].coverage[0].length,
+      zeros = Array(sites_in_partition).fill(0),
+      fasta = _.keys(branchAttributes[partition]).map(key => {
+        return {
+          header: key,
+          seq: branchAttributes[partition][key].codon[0].join("")
+        };
+      });
+    if(partition == 0) {
+      partition_data.offset = 0;
+    } else {
+      let previous_data = all_partition_data[partition-1];
+      partition_data.offset = previous_data.sites + previous_data.offset;
+    }
+    partition_data.sites = sites_in_partition;
+    partition_data.syn_substitutions = _.values(branchAttributes[partition])
+      .map(ba => ba["synonymous substitution count"][0])
+      .reduce((a, b) => a.map((d, i) => d + b[i]), zeros),
+    partition_data.nonsyn_substitutions = _.values(branchAttributes[partition])
+      .map(ba => ba["nonsynonymous substitution count"][0])
+      .reduce((a, b) => a.map((d, i) => d + b[i]), zeros);
+    const newick = trees[partition].newickString,
+      tree = new phylotreev1(newick);
+    placenodes(tree, true);
+    const number_of_links = tree.links.length,
+      tree_hash = _.object(
+        tree.links.map(l => l.target.data.name),
+        Array(number_of_links).fill(true)
+      );
+    tree_hash.root = false;
+    const node_to_ordered_index = _.object(
+      tree.node_order,
+      _.range(tree.node_order.length)
+    );
+    partition_data.tree = tree;
+    partition_data.sequence_data = fasta
+      .filter(record => tree_hash[record.header])
+      .sort((a, b) => {
+        const a_index = node_to_ordered_index[a.header],
+          b_index = node_to_ordered_index[b.header];
+        return a_index - b_index;
+      });
+    if (number_of_links != partition_data.sequence_data.length) {
+      throw "Mismatch in FASTA/tree... refusing to proceed!";
+    }
+    all_partition_data.push(partition_data);
+    partition_map.push(...Array(sites_in_partition).fill(partition));
+  }
+  return (<SLACAlignmentTreeCountWidget
+    partitionData={all_partition_data}
+    partitionMap={partition_map}
+    branchAttributes={branchAttributes}
+  />)
+}
+
+class SLACAlignmentTreeCountWidget extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      current_site: 1,
+      current_site_index: 0,
+      select_value: "amino-acid"
+    };
+  }
+  savePNG() {
+    saveSvgAsPng(
+      document.getElementById("slac-widget"),
+      "slac-ancestral-phylo.png"
+    );
+  }
+  handleInputChange(e) {
+    const new_current_site = e.target.value ? +e.target.value : "",
+      max_sites = this.props.partitionMap.length;
+    if (new_current_site == "") {
+      this.setState({
+        current_site: "",
+        current_site_index: 0
+      });
+    } else if (new_current_site > 0 && new_current_site <= max_sites) {
+      this.setState({
+        current_site: new_current_site,
+        current_site_index: new_current_site - 1
+      });
+    }
+  }
+  render() {
+    const { site_size, branchAttributes } = this.props,
+      { current_site, current_site_index } = this.state,
+      partition = this.props.partitionMap[current_site_index],
+      horizontal_pad = 10,
+      vertical_pad = site_size / 2,
+      {
+        sequence_data, tree, syn_substitutions, nonsyn_substitutions, offset
+      } = this.props.partitionData[partition],
+      offset_site_index = current_site_index - offset,
+      phylotree_props = {
+        width: 700 - 2 * horizontal_pad,
+        height: site_size * sequence_data.length - 2 * vertical_pad,
+        tree: tree,
+        transform: `translate(2, ${vertical_pad})`
+      },
+      site_padding = 5,
+      codon_label_height = 30,
+      syn_count = syn_substitutions[offset_site_index],
+      nonsyn_count = nonsyn_substitutions[offset_site_index],
+      total_count = syn_count + nonsyn_count,
+      column_padding = 40,
+      codon_column_width = 4 * site_size + site_padding + column_padding,
+      ccw_nopad = 4 * site_size + site_padding,
+      bar_height = 200,
+      svg_props = {
+        width: phylotree_props.width + codon_column_width + 2 * horizontal_pad,
+        height:
+          phylotree_props.height +
+          codon_label_height +
+          bar_height +
+          2 * vertical_pad
+      },
+      legend_colors = {
+        total: "black",
+        nonsyn: "#00a99d",
+        syn: "grey"
+      };
+    const bar_scale = d3.scale
+        .linear()
+        .domain([0, total_count])
+        .range([0, bar_height]),
+      bar_padding = 5,
+      bar_width = (ccw_nopad - 4 * bar_padding) / 3;
+
+    const { select_value } = this.state;
+    const codons = {},
+      codon_colors = d3.scale.category10().range();
+    var codon_color_index = 0;
+    tree.traverse_and_compute(node => {
+      if (node.data.name != "root") {
+        if (select_value != "none") {
+          const character =
+            branchAttributes[partition][node.data.name][select_value][0][
+              offset_site_index
+            ];
+          node.data.annotation = character;
+          if (select_value == "codon") {
+            if (!codons[character]) {
+              codons[character] = codon_colors[codon_color_index++];
+            }
+          }
+        } else {
+          node.data.annotation = null;
+        }
+      }
+    });
+
+    const color_scale =
+      select_value != "codon" ? colors.amino_acid_colors : codons;
+    return (
+      <div>
+        <div
+          style={{
+            display: "flex",
+            justifyContent: "space-around",
+            alignItems: "center",
+            paddingBottom: 20
+          }}
+        >
+          <span>
+            Current site:
+            <input
+              type="number"
+              value={current_site}
+              onChange={e => this.handleInputChange(e)}
+              ref={input => {
+                this.numberInput = input;
+              }}
+              style={{ width: 50 }}
+            />
+          </span>
+
+          <span>
+            Show{" "}
+            <select
+              value={this.state.select_value}
+              onChange={e => this.setState({ select_value: e.target.value })}
+            >
+              <option value="amino-acid">amino acid</option>
+              <option value="codon">codon</option>
+              <option value="none">none</option>
+            </select>{" "}
+            on tree
+          </span>
+
+          <button
+            onClick={() => this.savePNG()}
+            type="button"
+            className="btn btn-secondary"
+            data-dismiss="modal"
+          >
+            <span className="far fa-save" />
+            PNG
+          </button>
+        </div>
+
+        <div style={{ textAlign: "center" }}>
+          <svg
+            {...svg_props}
+            id="slac-widget"
+            style={{ fontFamily: "sans-serif" }}
+          >
+            <g transform="translate(0, 5)">
+              <rect
+                x={0}
+                y={0}
+                width={svg_props.width}
+                height={svg_props.height}
+                fill="white"
+              />
+              <g transform={`translate(${phylotree_props.width - 100}, 50)`}>
+                <rect
+                  x={5}
+                  y={10}
+                  width={20}
+                  height={20}
+                  fill={legend_colors.total}
+                />
+                <text x={0} y={20} textAnchor="end" alignmentBaseline="middle">
+                  Total
+                </text>
+                <rect
+                  x={5}
+                  y={40}
+                  width={20}
+                  height={20}
+                  fill={legend_colors.nonsyn}
+                />
+                <text x={0} y={50} textAnchor="end" alignmentBaseline="middle">
+                  Nonsynonymous
+                </text>
+                <rect
+                  x={5}
+                  y={70}
+                  width={20}
+                  height={20}
+                  fill={legend_colors.syn}
+                />
+                <text x={0} y={80} textAnchor="end" alignmentBaseline="middle">
+                  Synonymous
+                </text>
+              </g>
+              <SitePlotAxis
+                data={[syn_count, nonsyn_count, total_count]}
+                axis_label="Substitution counts"
+                height={bar_height}
+                label_width={75}
+                translateX={phylotree_props.width - 75}
+                padding={{ top: 0, right: 0, bottom: 0, left: 0 }}
+              />
+              <rect
+                x={phylotree_props.width + bar_padding}
+                y={bar_height - bar_scale(total_count)}
+                width={bar_width}
+                height={bar_scale(total_count)}
+                fill={legend_colors.total}
+              />
+              <rect
+                x={phylotree_props.width + 2 * bar_padding + bar_width}
+                y={bar_height - bar_scale(nonsyn_count)}
+                width={bar_width}
+                height={bar_scale(nonsyn_count)}
+                fill={legend_colors.nonsyn}
+              />
+              <rect
+                x={phylotree_props.width + 3 * bar_padding + 2 * bar_width}
+                y={bar_height - bar_scale(syn_count)}
+                width={bar_width}
+                height={bar_scale(syn_count)}
+                fill={legend_colors.syn}
+              />
+              <g transform={`translate(0, ${bar_height + 5})`}>
+                <Phylotree
+                  {...phylotree_props}
+                  internalNodeLabels
+                  skipPlacement
+                  highlightBranches={color_scale}
+                />
+                <CodonColumn
+                  site={offset_site_index}
+                  translateX={phylotree_props.width + horizontal_pad}
+                  site_size={site_size}
+                  site_padding={site_padding}
+                  codon_label_height={codon_label_height}
+                  height={phylotree_props.height}
+                  sequence_data={sequence_data}
+                  tree={tree}
+                  {...this.state}
+                />
+              </g>
+            </g>
+          </svg>
+        </div>
+      </div>
+    );
+  }
+}
+
+SLACAlignmentTreeCountWidget.defaultProps = {
+  site_size: 20
+};
+
 class SLACContents extends React.Component {
   constructor(props) {
     super(props);
@@ -1247,17 +1565,6 @@ class SLACContents extends React.Component {
 
     const { analysis_results, tree } = self.state;
     if (analysis_results) {
-      const branch_attributes = analysis_results["branch attributes"]["0"],
-        fasta = self.props.fasta,
-        zeros = d3.range(analysis_results.input["number of sites"]).map(n => 0),
-        syn_substitutions = _.values(branch_attributes)
-          .map(ba => ba["synonymous substitution count"][0])
-          .reduce((a, b) => a.map((d, i) => d + b[i]), zeros),
-        nonsyn_substitutions = _.values(branch_attributes)
-          .map(ba => ba["nonsynonymous substitution count"][0])
-          .reduce((a, b) => a.map((d, i) => d + b[i]), zeros),
-        newick = analysis_results.trees[0].newickString;
-
       var trees = analysis_results
         ? {
             newick: analysis_results.input.trees,
@@ -1427,12 +1734,8 @@ class SLACContents extends React.Component {
                 title="SLAC Phylogenetic Alignment"
                 popover="<ul><li>View individual codons and amino acids with corresponding phylogenetic information.</li><li>Select individual sites and view substitution counts across the phylogeny.</li></ul>"
               />
-              <AlignmentTree
-                fasta={fasta}
-                newick={newick}
-                syn_substitutions={syn_substitutions}
-                nonsyn_substitutions={nonsyn_substitutions}
-                branchAttributes={branch_attributes}
+              <SLACAlignmentTreeCountWrapper
+                json={analysis_results}
               />
             </div>
           </div>
@@ -1443,7 +1746,7 @@ class SLACContents extends React.Component {
   }
 }
 
-function SLAC(props) {
+export function SLAC(props) {
   return (
     <ResultsPage
       data={props.data}
@@ -1467,7 +1770,13 @@ function SLAC(props) {
 
 // Will need to make a call to this
 // omega distributions
-function render_slac(data, element, fasta, originalFile, analysisLog) {
+export default function render_slac(
+  data,
+  element,
+  fasta,
+  originalFile,
+  analysisLog
+) {
   ReactDOM.render(
     <SLAC
       data={data}
@@ -1478,5 +1787,3 @@ function render_slac(data, element, fasta, originalFile, analysisLog) {
     document.getElementById(element)
   );
 }
-
-export { SLAC, render_slac };
