@@ -17,7 +17,7 @@ module.exports = env => {
       historyApiFallback: true
     },
     output: {
-      path: path.resolve(__dirname, "public/"),
+      path: path.resolve(__dirname, "dist/"),
       filename: "[name].js",
       library: "hyphyVision"
     },
@@ -102,7 +102,7 @@ module.exports = env => {
       new PreloadWebpackPlugin(),
       new HtmlWebpackPlugin({
         title: "HyPhy Vision",
-        filename: path.resolve("public", "index.html")
+        filename: path.resolve("dist", "index.html")
       }),
       new MiniCssExtractPlugin({
         // Options similar to the same options in webpackOptions.output
